refactor common types into its own util class
