+++ src/urllib3/__init__.py
@@ -11,7 +11,8 @@ from typing import Any, Mapping, Optional, TextIO, Type, Union
 from . import exceptions
 from ._collections import HTTPHeaderDict
 from ._version import __version__
-from .connection import _TYPE_BODY
+
+from .util.typing import _TYPE_BODY
 from .connectionpool import HTTPConnectionPool, HTTPSConnectionPool, connection_from_url
 from .filepost import _TYPE_FIELDS, encode_multipart_formdata
 from .poolmanager import PoolManager, ProxyManager, proxy_from_url
