+++ src/urllib3/util/typing.py
@@ -1,0 +1,4 @@
+
+from typing import Union, IO, Any, Iterable
+
+_TYPE_BODY = Union[bytes, IO[Any], Iterable[bytes], str]
